+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [1.1.3] - 24/09/2019
+* Added Flutter driver reporter - the driver logs all messages (even non-error ones) to stderr and will cause the process to be marked as failed by a CI server becuase of this.  So this reporter redirect the messages to the appropiate output stream (stdout / stderr).
+* FlutterWorld - added missing `super.dispose()` call
+
 ## [1.1.2] - 22/09/2019
 * Fixed lint warnings